Write the first test, skip the rest
I played around with two different ideas. In the end, I went with this
approach, because it keeps my CLI from having to know about how to talk
to the input and output streams, so much simpler.

This implies that there is something (the binary, presumably) sitting
above it, feeding it each line of input and printing each line of
output. I will likely omit testing it as it will be very small and
clearly working, and more annoying to test. Though, if I did want to,
I'd stick it in a class that received its streams and just read
from/printed to them, then I'd use StringIO.