Some unit tests to help with refactoring, and confidence
Though, as I'm making these tests,
I notice that all my methods end in "command?". In fact, it was so clear
that they all belonged together that I even made a separate grouping of
tests for them.

So I'm actually thinking now that they should be separated from the rest
of this class.