Abstracted my test away from how I'm setting up my data
Notice this!

Have you ever made a test that says something like "it gets initialized with this value"
And then later you wanted to change the thing, but it was hard because your tests were
in the way?

Notice that I'm doing exactly the opposite! I'm staying as far away from these
implementation details as I can. I want these tests to live a long time,
and only break when the logic changes