Tests for the CsvReader
When I went to update it, I realized the default name
and my one test's name are the same. Also, wound up
with an edge case where I could explicitly pass nil,
and I'd want that to be turned into a default.

Thinking about how much work it would be to show that
it does the correct thing from that high-level integration test,
I decided it was time to make some unit tests