Add a Gemfile and FakeFS
The Gemfile locks us down to the gems we want to use
otherwise, we might come back next month, the gems that
get loaded are totally different, and none of our code
runs anymore!

FakeFS is a library that allows us to fake out the file system.
We'll pretend-write a file to it, and when our code goes to
read, it will find our pretend csv!