Revert "Formatter does the sorting"
This reverts commit 5797faf19d10b2198d47d83fce68a746d98100f3.

Got to thinking about it, and while it initially made sense to me that
the thing doing the formatting of the data would choose what order the
data should appear in, I realized that it can't ever have enough context
to really know how to sort the data. For example, right now its just by
a single attribute, but what if I wanted to sort by "distance from a bus stop"?
this would have no way to understand and consider that information.

So I'm leaving it up to the caller to determine what order to put the
data in, and this class will only deal with formatting the pre-oredered data.